# HIVE-24353: performance: do not throw exceptions when parsing dates