--- a/common/src/java/org/apache/hadoop/hive/common/type/TimestampTZUtil.java
+++ b/common/src/java/org/apache/hadoop/hive/common/type/TimestampTZUtil.java
@@ -34,6 +34,7 @@ import java.time.format.DateTimeParseException;
 import java.time.format.TextStyle;
 import java.time.temporal.ChronoField;
 import java.time.temporal.TemporalAccessor;
+import java.time.temporal.TemporalQueries;
 import java.util.TimeZone;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -71,33 +72,24 @@ public class TimestampTZUtil {
   public static TimestampTZ parse(String s, ZoneId defaultTimeZone) {
     // need to handle offset with single digital hour, see JDK-8066806
     s = handleSingleDigitHourOffset(s);
-    ZonedDateTime zonedDateTime;
-    try {
-      zonedDateTime = ZonedDateTime.parse(s, FORMATTER);
-    } catch (DateTimeParseException e) {
-      // try to be more tolerant
-      // if the input is invalid instead of incomplete, we'll hit exception here again
-      TemporalAccessor accessor = FORMATTER.parse(s);
-      // LocalDate must be present
-      LocalDate localDate = LocalDate.from(accessor);
-      LocalTime localTime;
-      ZoneId zoneId;
-      try {
-        localTime = LocalTime.from(accessor);
-      } catch (DateTimeException e1) {
-        localTime = DEFAULT_LOCAL_TIME;
-      }
-      try {
-        zoneId = ZoneId.from(accessor);
-      } catch (DateTimeException e2) {
-        if (defaultTimeZone == null) {
-          throw new DateTimeException("Time Zone not available");
-        }
-        zoneId = defaultTimeZone;
-      }
-      zonedDateTime = ZonedDateTime.of(localDate, localTime, zoneId);
+    TemporalAccessor accessor = FORMATTER.parse(s);
+
+    LocalDate localDate = accessor.query(TemporalQueries.localDate());
+
+    LocalTime localTime = accessor.query(TemporalQueries.localTime());
+    if (localTime == null) {
+      localTime = DEFAULT_LOCAL_TIME;
     }
 
+    ZoneId zoneId = accessor.query(TemporalQueries.zone());
+    if (zoneId == null) {
+      zoneId = defaultTimeZone;
+      if (zoneId == null) {
+        throw new DateTimeException("Time Zone not available");
+      }
+    }
+
+    ZonedDateTime zonedDateTime = ZonedDateTime.of(localDate, localTime, zoneId);
     if (defaultTimeZone == null) {
       return new TimestampTZ(zonedDateTime);
     }