This resolves https://issues.apache.org/jira/browse/HIVE-24353

### What changes were proposed in this pull request?
Change timestampTZ parsing from using convenience methods that throw exceptions to a chain of queries against the parsed temporalAccessor. 

### Why are the changes needed?
This improves the performance for timestamp parsing. 

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests are covering all branches already. 
One test was added, but it has no assertions.