@VPriesnitz 2021-01-28
@abstractdog  The average time spent over 10 run: 
Before: 225.5 ms
After:  48.7 ms 

The timing gets worse with deeper stacks. When adding 100 more method invocations to the stack the difference becomes more apparent:
Before: 494.8 ms
After:  54.2 ms 

I think that in hive, the latter is a more realistic stack depth.