@belugabehr 2021-01-28
Lemme take a look.  @abstractdog as it relates to HIVE-24693, this is not necessary, as HIVE-24693 removes this parsing from the code path altogether.